refactor: update pytest configuration and add utility classes for testing
