+++ pytest.ini
@@ -2,3 +2,4 @@
 testpaths = test
 python_files = *_test.py
 addopts = -v --tb=short
+pythonpath = /workspaces/InfraestructuraDiligencia
