+++ test/data/daos/topic_dao_test.py
@@ -6,24 +6,7 @@ from digiliencia.exc.dao_create_exc import DAOCreateError
 from digiliencia.exc.dao_delete_exc import DAODeleteError
 from digiliencia.exc.dao_read_exc import DAOReadError
 from digiliencia.exc.dao_update_exc import DAOUpdateError
-
-
-class DummyResult:
-    def single(self):
-        return None
-
-
-class DummySession:
-    def run(self, *args, **kwargs):
-        return DummyResult()
-
-
-class DummyContextManager:
-    def __enter__(self):
-        return DummySession()
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        return None
+from test.data.utils import DummyContextManager
 
 
 class TestTopicDAO:
